Moved classes to new MVC packages
