+++ Team18Project/app/Doc/com/example/team18project/AllQRCodesFragment.html
@@ -125,7 +125,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>androidx.fragment.app.Fragment</li>
 <li>
 <ul class="inheritance">
-<li>com.example.team18project.AllQRCodesFragment</li>
+<li>com.example.team18project.view.AllQRCodesFragment</li>
 </ul>
 </li>
 </ul>
