+++ Team18Project/app/Doc/com/example/team18project/Comment.html
@@ -122,7 +122,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>com.example.team18project.Comment</li>
+<li>com.example.team18project.model.Comment</li>
 </ul>
 </li>
 </ul>
