+++ Team18Project/app/Doc/com/example/team18project/HomeFragment.html
@@ -125,7 +125,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>androidx.fragment.app.Fragment</li>
 <li>
 <ul class="inheritance">
-<li>com.example.team18project.HomeFragment</li>
+<li>com.example.team18project.view.HomeFragment</li>
 </ul>
 </li>
 </ul>
@@ -230,7 +230,7 @@ extends androidx.fragment.app.Fragment</pre>
 </tr>
 <tr id="i0" class="altColor">
 <td class="colFirst"><code>static <a href="HomeFragment.html" title="class in com.example.team18project">HomeFragment</a></code></td>
-<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="#newInstance(com.example.team18project.Player)">newInstance</a></span>&#8203;(<a href="Player.html" title="class in com.example.team18project">Player</a>&nbsp;player)</code></th>
+<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="#newInstance(com.example.team18project.model.Player)">newInstance</a></span>&#8203;(<a href="Player.html" title="class in com.example.team18project">Player</a>&nbsp;player)</code></th>
 <td class="colLast">
 <div class="block">Use this factory method to create a new instance of
  this fragment using the provided parameters.</div>
@@ -348,7 +348,7 @@ extends androidx.fragment.app.Fragment</pre>
 <!--   -->
 </a>
 <h3>Method Detail</h3>
-<a id="newInstance(com.example.team18project.Player)">
+<a id="newInstance(com.example.team18project.model.Player)">
 <!--   -->
 </a>
 <ul class="blockList">
