+++ Team18Project/app/Doc/com/example/team18project/Player.html
@@ -122,7 +122,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>com.example.team18project.Player</li>
+<li>com.example.team18project.model.Player</li>
 </ul>
 </li>
 </ul>
@@ -140,7 +140,7 @@ implements android.os.Parcelable, java.io.Serializable</pre>
 <div class="block">Class for modelling players. Stores scanned QR codes and account information.</div>
 <dl>
 <dt><span class="seeLabel">See Also:</span></dt>
-<dd><a href="../../../serialized-form.html#com.example.team18project.Player">Serialized Form</a></dd>
+<dd><a href="../../../serialized-form.html#com.example.team18project.model.Player">Serialized Form</a></dd>
 </dl>
 </li>
 </ul>
@@ -255,7 +255,7 @@ implements android.os.Parcelable, java.io.Serializable</pre>
 </tr>
 <tr id="i0" class="altColor">
 <td class="colFirst"><code>void</code></td>
-<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="#addQRCode(com.example.team18project.QRCode)">addQRCode</a></span>&#8203;(<a href="QRCode.html" title="class in com.example.team18project">QRCode</a>&nbsp;qrCode)</code></th>
+<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="#addQRCode(com.example.team18project.model.QRCode)">addQRCode</a></span>&#8203;(<a href="QRCode.html" title="class in com.example.team18project">QRCode</a>&nbsp;qrCode)</code></th>
 <td class="colLast">
 <div class="block">This method adds a QRCode object to the list of codes and Firestore database.</div>
 </td>
@@ -323,7 +323,7 @@ implements android.os.Parcelable, java.io.Serializable</pre>
 </tr>
 <tr id="i10" class="altColor">
 <td class="colFirst"><code>void</code></td>
-<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="#removeQRCode(com.example.team18project.QRCode)">removeQRCode</a></span>&#8203;(<a href="QRCode.html" title="class in com.example.team18project">QRCode</a>&nbsp;qrCode)</code></th>
+<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="#removeQRCode(com.example.team18project.model.QRCode)">removeQRCode</a></span>&#8203;(<a href="QRCode.html" title="class in com.example.team18project">QRCode</a>&nbsp;qrCode)</code></th>
 <td class="colLast">
 <div class="block">This method removes a QRCode object from the list of codes and Firestore database.</div>
 </td>
@@ -502,7 +502,7 @@ implements android.os.Parcelable, java.io.Serializable</pre>
 <!--   -->
 </a>
 <h3>Method Detail</h3>
-<a id="addQRCode(com.example.team18project.QRCode)">
+<a id="addQRCode(com.example.team18project.model.QRCode)">
 <!--   -->
 </a>
 <ul class="blockList">
@@ -516,7 +516,7 @@ implements android.os.Parcelable, java.io.Serializable</pre>
 </dl>
 </li>
 </ul>
-<a id="removeQRCode(com.example.team18project.QRCode)">
+<a id="removeQRCode(com.example.team18project.model.QRCode)">
 <!--   -->
 </a>
 <ul class="blockList">
