+++ Team18Project/app/Doc/com/example/team18project/QRArrayAdapter.html
@@ -128,7 +128,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>android.widget.ArrayAdapter&lt;<a href="QRCode.html" title="class in com.example.team18project">QRCode</a>&gt;</li>
 <li>
 <ul class="inheritance">
-<li>com.example.team18project.QRArrayAdapter</li>
+<li>com.example.team18project.model.QRArrayAdapter</li>
 </ul>
 </li>
 </ul>
