+++ Team18Project/app/Doc/com/example/team18project/QRCode.html
@@ -122,7 +122,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>com.example.team18project.QRCode</li>
+<li>com.example.team18project.model.QRCode</li>
 </ul>
 </li>
 </ul>
