+++ Team18Project/app/Doc/com/example/team18project/QRMenuFragment.html
@@ -128,7 +128,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>androidx.fragment.app.DialogFragment</li>
 <li>
 <ul class="inheritance">
-<li>com.example.team18project.QRMenuFragment</li>
+<li>com.example.team18project.view.QRMenuFragment</li>
 </ul>
 </li>
 </ul>
@@ -201,7 +201,7 @@ extends androidx.fragment.app.DialogFragment</pre>
 <th class="colLast" scope="col">Description</th>
 </tr>
 <tr class="altColor">
-<th class="colConstructorName" scope="row"><code><span class="memberNameLink"><a href="#%3Cinit%3E(com.example.team18project.Player,com.example.team18project.QRCode,com.example.team18project.QRArrayAdapter)">QRMenuFragment</a></span>&#8203;(<a href="Player.html" title="class in com.example.team18project">Player</a>&nbsp;player,
+<th class="colConstructorName" scope="row"><code><span class="memberNameLink"><a href="#%3Cinit%3E(com.example.team18project.model.Player,com.example.team18project.model.QRCode,com.example.team18project.model.QRArrayAdapter)">QRMenuFragment</a></span>&#8203;(<a href="Player.html" title="class in com.example.team18project">Player</a>&nbsp;player,
               <a href="QRCode.html" title="class in com.example.team18project">QRCode</a>&nbsp;code,
               <a href="QRArrayAdapter.html" title="class in com.example.team18project">QRArrayAdapter</a>&nbsp;adapter)</code></th>
 <td class="colLast">&nbsp;</td>
@@ -281,7 +281,7 @@ extends androidx.fragment.app.DialogFragment</pre>
 <!--   -->
 </a>
 <h3>Constructor Detail</h3>
-<a id="&lt;init&gt;(com.example.team18project.Player,com.example.team18project.QRCode,com.example.team18project.QRArrayAdapter)">
+<a id="&lt;init&gt;(com.example.team18project.model.Player,com.example.team18project.model.QRCode,com.example.team18project.model.QRArrayAdapter)">
 <!--   -->
 </a>
 <ul class="blockListLast">
