+++ Team18Project/app/Doc/com/example/team18project/ScanQRCode.html
@@ -146,7 +146,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>androidx.appcompat.app.AppCompatActivity</li>
 <li>
 <ul class="inheritance">
-<li>com.example.team18project.ScanQRCode</li>
+<li>com.example.team18project.view.ScanQRCode</li>
 </ul>
 </li>
 </ul>
