+++ Team18Project/app/Doc/com/example/team18project/ScanQRCodeFragment.html
@@ -143,7 +143,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 extends androidx.fragment.app.Fragment</pre>
 <div class="block">Fragment class that does the same thing as ScanQRCode. Left unused at the moment.
  A simple <code>Fragment</code> subclass.
- Use the <a href="#newInstance(com.example.team18project.Player)"><code>newInstance(com.example.team18project.Player)</code></a> factory method to
+ Use the <a href="#newInstance(com.example.team18project.model.Player)"><code>newInstance(com.example.team18project.model.Player)</code></a> factory method to
  create an instance of this fragment.</div>
 </li>
 </ul>
@@ -205,7 +205,7 @@ extends androidx.fragment.app.Fragment</pre>
 </tr>
 <tr id="i0" class="altColor">
 <td class="colFirst"><code>static <a href="ScanQRCodeFragment.html" title="class in com.example.team18project">ScanQRCodeFragment</a></code></td>
-<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="#newInstance(com.example.team18project.Player)">newInstance</a></span>&#8203;(<a href="Player.html" title="class in com.example.team18project">Player</a>&nbsp;player)</code></th>
+<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="#newInstance(com.example.team18project.model.Player)">newInstance</a></span>&#8203;(<a href="Player.html" title="class in com.example.team18project">Player</a>&nbsp;player)</code></th>
 <td class="colLast">&nbsp;</td>
 </tr>
 <tr id="i1" class="rowColor">
@@ -296,7 +296,7 @@ extends androidx.fragment.app.Fragment</pre>
 <!--   -->
 </a>
 <h3>Method Detail</h3>
-<a id="newInstance(com.example.team18project.Player)">
+<a id="newInstance(com.example.team18project.model.Player)">
 <!--   -->
 </a>
 <ul class="blockList">
