+++ Team18Project/app/Doc/com/example/team18project/SearchFragment.html
@@ -125,7 +125,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>androidx.fragment.app.Fragment</li>
 <li>
 <ul class="inheritance">
-<li>com.example.team18project.SearchFragment</li>
+<li>com.example.team18project.view.SearchFragment</li>
 </ul>
 </li>
 </ul>
@@ -217,7 +217,7 @@ extends androidx.fragment.app.Fragment</pre>
 </tr>
 <tr id="i0" class="altColor">
 <td class="colFirst"><code>void</code></td>
-<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="#generateUserList(com.example.team18project.SearchFragment.UserListCallback)">generateUserList</a></span>&#8203;(<a href="SearchFragment.UserListCallback.html" title="interface in com.example.team18project">SearchFragment.UserListCallback</a>&nbsp;callback)</code></th>
+<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="#generateUserList(com.example.team18project.view.SearchFragment.UserListCallback)">generateUserList</a></span>&#8203;(<a href="SearchFragment.UserListCallback.html" title="interface in com.example.team18project">SearchFragment.UserListCallback</a>&nbsp;callback)</code></th>
 <td class="colLast">&nbsp;</td>
 </tr>
 <tr id="i1" class="rowColor">
@@ -340,7 +340,7 @@ extends androidx.fragment.app.Fragment</pre>
 </dl>
 </li>
 </ul>
-<a id="generateUserList(com.example.team18project.SearchFragment.UserListCallback)">
+<a id="generateUserList(com.example.team18project.view.SearchFragment.UserListCallback)">
 <!--   -->
 </a>
 <ul class="blockListLast">
