+++ Team18Project/app/Doc/com/example/team18project/StatsFragment.html
@@ -125,7 +125,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>androidx.fragment.app.Fragment</li>
 <li>
 <ul class="inheritance">
-<li>com.example.team18project.StatsFragment</li>
+<li>com.example.team18project.view.StatsFragment</li>
 </ul>
 </li>
 </ul>
@@ -142,7 +142,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <pre>public class <span class="typeNameLabel">StatsFragment</span>
 extends androidx.fragment.app.Fragment</pre>
 <div class="block">A simple <code>Fragment</code> subclass.
- Use the <a href="#newInstance(com.example.team18project.Player)"><code>newInstance(com.example.team18project.Player)</code></a> factory method to
+ Use the <a href="#newInstance(com.example.team18project.model.Player)"><code>newInstance(com.example.team18project.model.Player)</code></a> factory method to
  create an instance of this fragment.</div>
 </li>
 </ul>
@@ -204,7 +204,7 @@ extends androidx.fragment.app.Fragment</pre>
 </tr>
 <tr id="i0" class="altColor">
 <td class="colFirst"><code>static <a href="StatsFragment.html" title="class in com.example.team18project">StatsFragment</a></code></td>
-<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="#newInstance(com.example.team18project.Player)">newInstance</a></span>&#8203;(<a href="Player.html" title="class in com.example.team18project">Player</a>&nbsp;player)</code></th>
+<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="#newInstance(com.example.team18project.model.Player)">newInstance</a></span>&#8203;(<a href="Player.html" title="class in com.example.team18project">Player</a>&nbsp;player)</code></th>
 <td class="colLast">
 <div class="block">Use this factory method to create a new instance of
  this fragment using the provided parameters.</div>
@@ -279,7 +279,7 @@ extends androidx.fragment.app.Fragment</pre>
 <!--   -->
 </a>
 <h3>Method Detail</h3>
-<a id="newInstance(com.example.team18project.Player)">
+<a id="newInstance(com.example.team18project.model.Player)">
 <!--   -->
 </a>
 <ul class="blockList">
