+++ Team18Project/app/Doc/index-files/index-1.html
@@ -98,7 +98,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <dl>
 <dt><span class="memberNameLink"><a href="../com/example/team18project/ExampleUnitTest.html#addition_isCorrect()">addition_isCorrect()</a></span> - Method in class com.example.team18project.<a href="../com/example/team18project/ExampleUnitTest.html" title="class in com.example.team18project">ExampleUnitTest</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="../com/example/team18project/Player.html#addQRCode(com.example.team18project.QRCode)">addQRCode(QRCode)</a></span> - Method in class com.example.team18project.<a href="../com/example/team18project/Player.html" title="class in com.example.team18project">Player</a></dt>
+<dt><span class="memberNameLink"><a href="../com/example/team18project/Player.html#addQRCode(com.example.team18project.model.QRCode)">addQRCode(QRCode)</a></span> - Method in class com.example.team18project.<a href="../com/example/team18project/Player.html" title="class in com.example.team18project">Player</a></dt>
 <dd>
 <div class="block">This method adds a QRCode object to the list of codes and Firestore database.</div>
 </dd>
