+++ Team18Project/app/Doc/index-files/index-10.html
@@ -105,19 +105,19 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <div class="block">Use this factory method to create a new instance of
  this fragment using the provided parameters.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../com/example/team18project/HomeFragment.html#newInstance(com.example.team18project.Player)">newInstance(Player)</a></span> - Static method in class com.example.team18project.<a href="../com/example/team18project/HomeFragment.html" title="class in com.example.team18project">HomeFragment</a></dt>
+<dt><span class="memberNameLink"><a href="../com/example/team18project/HomeFragment.html#newInstance(com.example.team18project.model.Player)">newInstance(Player)</a></span> - Static method in class com.example.team18project.<a href="../com/example/team18project/HomeFragment.html" title="class in com.example.team18project">HomeFragment</a></dt>
 <dd>
 <div class="block">Use this factory method to create a new instance of
  this fragment using the provided parameters.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../com/example/team18project/ProfileFragment.html#newInstance(com.example.team18project.Player)">newInstance(Player)</a></span> - Static method in class com.example.team18project.<a href="../com/example/team18project/ProfileFragment.html" title="class in com.example.team18project">ProfileFragment</a></dt>
+<dt><span class="memberNameLink"><a href="../com/example/team18project/ProfileFragment.html#newInstance(com.example.team18project.model.Player)">newInstance(Player)</a></span> - Static method in class com.example.team18project.<a href="../com/example/team18project/ProfileFragment.html" title="class in com.example.team18project">ProfileFragment</a></dt>
 <dd>
 <div class="block">Use this factory method to create a new instance of
  this fragment using the provided parameters.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../com/example/team18project/ScanQRCodeFragment.html#newInstance(com.example.team18project.Player)">newInstance(Player)</a></span> - Static method in class com.example.team18project.<a href="../com/example/team18project/ScanQRCodeFragment.html" title="class in com.example.team18project">ScanQRCodeFragment</a></dt>
+<dt><span class="memberNameLink"><a href="../com/example/team18project/ScanQRCodeFragment.html#newInstance(com.example.team18project.model.Player)">newInstance(Player)</a></span> - Static method in class com.example.team18project.<a href="../com/example/team18project/ScanQRCodeFragment.html" title="class in com.example.team18project">ScanQRCodeFragment</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="../com/example/team18project/StatsFragment.html#newInstance(com.example.team18project.Player)">newInstance(Player)</a></span> - Static method in class com.example.team18project.<a href="../com/example/team18project/StatsFragment.html" title="class in com.example.team18project">StatsFragment</a></dt>
+<dt><span class="memberNameLink"><a href="../com/example/team18project/StatsFragment.html#newInstance(com.example.team18project.model.Player)">newInstance(Player)</a></span> - Static method in class com.example.team18project.<a href="../com/example/team18project/StatsFragment.html" title="class in com.example.team18project">StatsFragment</a></dt>
 <dd>
 <div class="block">Use this factory method to create a new instance of
  this fragment using the provided parameters.</div>
