+++ Team18Project/app/Doc/index-files/index-13.html
@@ -124,7 +124,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <dd>
 <div class="block">Fragment for menu that pops up when a QR Code is clicked</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../com/example/team18project/QRMenuFragment.html#%3Cinit%3E(com.example.team18project.Player,com.example.team18project.QRCode,com.example.team18project.QRArrayAdapter)">QRMenuFragment(Player, QRCode, QRArrayAdapter)</a></span> - Constructor for class com.example.team18project.<a href="../com/example/team18project/QRMenuFragment.html" title="class in com.example.team18project">QRMenuFragment</a></dt>
+<dt><span class="memberNameLink"><a href="../com/example/team18project/QRMenuFragment.html#%3Cinit%3E(com.example.team18project.model.Player,com.example.team18project.model.QRCode,com.example.team18project.model.QRArrayAdapter)">QRMenuFragment(Player, QRCode, QRArrayAdapter)</a></span> - Constructor for class com.example.team18project.<a href="../com/example/team18project/QRMenuFragment.html" title="class in com.example.team18project">QRMenuFragment</a></dt>
 <dd>&nbsp;</dd>
 </dl>
 <a href="index-1.html">A</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">D</a>&nbsp;<a href="index-4.html">E</a>&nbsp;<a href="index-5.html">G</a>&nbsp;<a href="index-6.html">H</a>&nbsp;<a href="index-7.html">I</a>&nbsp;<a href="index-8.html">L</a>&nbsp;<a href="index-9.html">M</a>&nbsp;<a href="index-10.html">N</a>&nbsp;<a href="index-11.html">O</a>&nbsp;<a href="index-12.html">P</a>&nbsp;<a href="index-13.html">Q</a>&nbsp;<a href="index-14.html">R</a>&nbsp;<a href="index-15.html">S</a>&nbsp;<a href="index-16.html">T</a>&nbsp;<a href="index-17.html">U</a>&nbsp;<a href="index-18.html">W</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes</a>&nbsp;<a href="../allpackages-index.html">All&nbsp;Packages</a></div>
