+++ Team18Project/app/Doc/index-files/index-14.html
@@ -96,7 +96,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </a>
 <h2 class="title">R</h2>
 <dl>
-<dt><span class="memberNameLink"><a href="../com/example/team18project/Player.html#removeQRCode(com.example.team18project.QRCode)">removeQRCode(QRCode)</a></span> - Method in class com.example.team18project.<a href="../com/example/team18project/Player.html" title="class in com.example.team18project">Player</a></dt>
+<dt><span class="memberNameLink"><a href="../com/example/team18project/Player.html#removeQRCode(com.example.team18project.model.QRCode)">removeQRCode(QRCode)</a></span> - Method in class com.example.team18project.<a href="../com/example/team18project/Player.html" title="class in com.example.team18project">Player</a></dt>
 <dd>
 <div class="block">This method removes a QRCode object from the list of codes and Firestore database.</div>
 </dd>
