+++ Team18Project/app/Doc/index-files/index-5.html
@@ -96,7 +96,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </a>
 <h2 class="title">G</h2>
 <dl>
-<dt><span class="memberNameLink"><a href="../com/example/team18project/SearchFragment.html#generateUserList(com.example.team18project.SearchFragment.UserListCallback)">generateUserList(SearchFragment.UserListCallback)</a></span> - Method in class com.example.team18project.<a href="../com/example/team18project/SearchFragment.html" title="class in com.example.team18project">SearchFragment</a></dt>
+<dt><span class="memberNameLink"><a href="../com/example/team18project/SearchFragment.html#generateUserList(com.example.team18project.view.SearchFragment.UserListCallback)">generateUserList(SearchFragment.UserListCallback)</a></span> - Method in class com.example.team18project.<a href="../com/example/team18project/SearchFragment.html" title="class in com.example.team18project">SearchFragment</a></dt>
 <dd>&nbsp;</dd>
 <dt><span class="memberNameLink"><a href="../com/example/team18project/Comment.html#getCid()">getCid()</a></span> - Method in class com.example.team18project.<a href="../com/example/team18project/Comment.html" title="class in com.example.team18project">Comment</a></dt>
 <dd>
