+++ Team18Project/app/Doc/serialized-form.html
@@ -100,10 +100,10 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <section role="region">
 <h2 title="Package">Package&nbsp;com.example.team18project</h2>
 <ul class="blockList">
-<li class="blockList"><a id="com.example.team18project.Player">
+<li class="blockList"><a id="com.example.team18project.model.Player">
 <!--   -->
 </a>
-<h3>Class <a href="com/example/team18project/Player.html" title="class in com.example.team18project">com.example.team18project.Player</a> extends java.lang.Object implements Serializable</h3>
+<h3>Class <a href="com/example/team18project/Player.html" title="class in com.example.team18project">com.example.team18project.model.Player</a> extends java.lang.Object implements Serializable</h3>
 <ul class="blockList">
 <li class="blockList">
 <h3>Serialized Fields</h3>
