+++ Team18Project/app/src/androidTest/java/com/example/team18project/SearchTesting.java
@@ -8,14 +8,11 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
 import android.content.Intent;
-import android.widget.EditText;
 
 import androidx.test.platform.app.InstrumentationRegistry;
 import androidx.test.rule.ActivityTestRule;
 
-import com.google.firebase.firestore.CollectionReference;
-import com.google.firebase.firestore.DocumentReference;
-import com.google.firebase.firestore.FirebaseFirestore;
+import com.example.team18project.view.MainActivity;
 import com.robotium.solo.Solo;
 
 import org.junit.After;
