+++ Team18Project/app/src/main/AndroidManifest.xml
@@ -19,10 +19,10 @@
         android:theme="@style/Theme.Team18Project"
         tools:targetApi="31">
         <activity
-            android:name=".ScanQRCode"
+            android:name=".view.ScanQRCode"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name=".view.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
