+++ Team18Project/app/src/main/java/com/example/team18project/OtherQRCodeFragment.java
@@ -1,19 +1,15 @@
 package com.example.team18project;
 
 import android.os.Bundle;
-import android.util.Log;
 import android.widget.ListView;
 
-import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.OnSuccessListener;
-import com.google.android.gms.tasks.Task;
+import com.example.team18project.model.Comment;
+import com.example.team18project.model.QRArrayAdapter;
+import com.example.team18project.model.QRCode;
 import com.google.firebase.firestore.CollectionReference;
-import com.google.firebase.firestore.DocumentReference;
-import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.EventListener;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.FirebaseFirestoreException;
@@ -21,9 +17,6 @@ import com.google.firebase.firestore.QueryDocumentSnapshot;
 import com.google.firebase.firestore.QuerySnapshot;
 
 import java.util.ArrayList;
-
-import java.io.File;
-
 
 
 public class OtherQRCodeFragment extends Fragment {
