+++ Team18Project/app/src/main/java/com/example/team18project/model/CommentArrayAdapter.java
@@ -1,15 +1,17 @@
-package com.example.team18project;
+package com.example.team18project.model;
 
 import android.content.Context;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
-import android.widget.ImageView;
 import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+
+import com.example.team18project.R;
+import com.example.team18project.model.Comment;
 
 import java.util.ArrayList;
