+++ Team18Project/app/src/main/java/com/example/team18project/view/ProfileFragment.java
@@ -1,7 +1,6 @@
-package com.example.team18project;
+package com.example.team18project.view;
 
 import android.os.Bundle;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -11,19 +10,12 @@ import android.widget.EditText;
 import android.widget.Switch;
 import android.widget.Toast;
 
-import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.Task;
-import com.google.firebase.firestore.CollectionReference;
-import com.google.firebase.firestore.DocumentReference;
-import com.google.firebase.firestore.FirebaseFirestore;
-import com.google.firebase.firestore.Query;
-import com.google.firebase.firestore.QuerySnapshot;
-
-import java.util.Arrays;
-import java.util.Collections;
+import com.example.team18project.controller.FirebaseWriter;
+import com.example.team18project.R;
+import com.example.team18project.controller.ProfileController;
+import com.example.team18project.model.Player;
 
 /**
  * A fragment used to represent the profile screen, where account information is displayed
