+++ Team18Project/app/src/main/java/com/example/team18project/view/StatsFragment.java
@@ -1,4 +1,4 @@
-package com.example.team18project;
+package com.example.team18project.view;
 
 import android.content.Context;
 import android.graphics.Bitmap;
@@ -13,6 +13,10 @@ import android.widget.ImageView;
 import android.widget.TextView;
 
 import androidx.fragment.app.Fragment;
+
+import com.example.team18project.R;
+import com.example.team18project.model.Player;
+import com.example.team18project.model.QRCode;
 
 /**
  * A simple {@link Fragment} subclass.
