+++ Team18Project/app/src/main/res/layout/fragment_all_qr_codes.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AllQRCodesFragment">
+    tools:context=".view.AllQRCodesFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
