+++ Team18Project/app/src/main/res/layout/fragment_profile.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frameLayout2"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ProfileFragment">
+    tools:context=".view.ProfileFragment">
 
     <TextView
         android:id="@+id/Account_Info_textView"
