+++ Team18Project/app/src/main/res/layout/fragment_qr_menu.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".QRMenuFragment">
+    tools:context=".view.QRMenuFragment">
 
     <TextView
         android:id="@+id/view_button"
