+++ Team18Project/app/src/main/res/layout/fragment_qr_view.xml
@@ -7,7 +7,7 @@
     android:fillViewport="true"
     android:orientation="vertical"
     android:overScrollMode="always"
-    tools:context=".QRViewFragment">
+    tools:context=".view.QRViewFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
