+++ Team18Project/app/src/main/res/layout/fragment_scan_q_r_code.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ScanQRCode">
+    tools:context=".view.ScanQRCode">
 
     <Switch
         android:id="@+id/take_photo_switch"
